pyproject.toml
[project]
name = "llm-council-plus"
version = "0.1.0"
description = "LLM Council Plus - Your Advanced AI Council"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
    "pydantic>=2.6.1",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "duckduckgo-search>=5.0.0",
    "beautifulsoup4>=4.12.3",
    "yake>=0.4.8",
    
    # Database
    "sqlalchemy>=2.0.27",
    "psycopg2-binary>=2.9.9",
    "pymysql>=1.1.0",
    
    # Tools
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "wikipedia>=1.4.0",
    "arxiv>=2.1.0",
    "yfinance>=0.2.36",
    
    # Documents
    "pypdf>=4.0.1",
    "python-docx>=1.1.0",
    "python-pptx>=0.6.23",
]
